cmd/govim: move ftplugin callback into separate file
Makes linking to these defaults easier
